+++ controlplane/src/types/index.ts
@@ -43,6 +43,12 @@ export interface SubgraphDTO {
   labels: Label[];
 }
 
+export interface MigrationSubgraph {
+  name: string;
+  routingURL: string;
+  schema: string;
+}
+
 export interface SchemaCheckDTO {
   id: string;
   targetID: string;
@@ -88,6 +94,7 @@ export interface APIKeyDTO {
   lastUsedAt: string;
   expiresAt: string;
   createdBy: string;
+  creatorUserID: string;
 }
 
 export interface ResponseMessage {