+++ keycloak/CHANGELOG.md
@@ -1,7 +1,20 @@
 # Change Log
+Images can be found [here](https://github.com/orgs/wundergraph/packages?repo_name=cosmo)
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.6.1](https://github.com/wundergraph/cosmo/compare/keycloak@0.6.0...keycloak@0.6.1) (2023-08-29)
+
+### Bug Fixes
+
+* improvements ([#18](https://github.com/wundergraph/cosmo/issues/18)) ([fdf2b29](https://github.com/wundergraph/cosmo/commit/fdf2b290ec57e502d8011e29e06a067d32afdf18)) (@JivusAyrus)
+
+# [0.6.0](https://github.com/wundergraph/cosmo/compare/keycloak@0.5.0...keycloak@0.6.0) (2023-08-28)
+
+### Features
+
+* improve keycloak accept invite ui ([#14](https://github.com/wundergraph/cosmo/issues/14)) ([4574b81](https://github.com/wundergraph/cosmo/commit/4574b815be2b8b03f87fc2844ca8c8abddf93f17)) (@JivusAyrus)
 
 # 0.5.0 (2023-08-24)